uKernel: Refactor main kernel to C++
The kernel is now written in C++! This might be a bad idea, but let's
roll with it. Some modules are still written in C, but those will be
rewritten some day

Moved asm to separate asm/ folder in the arch dir

Renamed "ps2" module to be more reflective of what it does, i.e. driver
for the i8042 controller