uKernel: Interrupts
This implements IRQ handlers for the 15 interrupt lines.
The PIC is remapped to offset the interrupts by 0x20 and then IDTR
is set up.

The interrupt handlers are defined in irqhandlers.cpp (should probably
do them in C though, as C++ might not be the best choice for handling
interrupts).