Build: More build system rework
Completely changed the build Makefiles for the kernel. The kernel Makefile
now adds objects from Makefile.objs files from specified folders in the
main makefile. This makes it so sources are properly organized, as
opposed to throwing everything into the arch folder.
Adding a new submodule is as easy as copying an example Makefile.objs file
into it and adding a line to Makefile.kernel.
In the future, it might be a good idea to implement the ability for
submodule makefiles to take over the compilation step for certain files
and add them to the final kernel link list.
Also, deleted some old code like cpuid things, which are pretty much useless
for now.