uKernel: Added ksleep
This is a sleep() implementation for the kernel. It sleeps for the
specified amount of milliseconds (approximately). Busy looping is not
exactly the best solution to this, in the future it might be good to add
scheduling to the timer, so the CPU is halted and then resumed by the timer
when the scheduled time has passed.