uKernel: PCI device probing
The kernel now probes the PCI buses and searches for valid PCI devices.
All valid devices are saved in a vector which can be queried by other
kernel modules. In the future it might be a good idea to add a "Query
by class" function to ease the work device drivers have to do to get
devices they are capable of serving.