uKernel/kmalloc: Early KMalloc initialization
KMalloc was previously initialized after parsing the multiboot structure,
which is very late in the boot process and does not allow us to dynamically
allocate memory which will be useful when processing the structure. KMalloc's
memory area is now defined in the kernel linker script instead, which
allows for early initialization (and gets rid of the overengineered approach
used previously).