uKernel: Refactor paging initialization and management
Removed the old paging code and replaced it with something relatively
better. We're no longer statically defining page tables for all
page directories, instead page tables will be added on-demand in the
kernel, and for initialization purposes, some space after the kernel
executable is reserved for pages (enough to map the entire binary).

Also, broke BootConfig.