LibGeneric: Strings / StringViews
This implements a string class for use by userland and the kernel,
and a StringView class to represent chunks of a given string. You can
add other strings, C strings and characters to the strings. They also
have a find() method implemented for finding a specified pattern.
The strings use amortized array allocation to speed up manipulation
of larger strings.