Scheduler: Create idle tasks in separate processes
To avoid premature optimization and having to maintain yet another
static process (like kerneld and userland init) in the executable,
idle tasks are now created in separate processes. This is a bit wasteful,
but it'll do for now. As a result of this though, you can name each
process with the APIC ID of the AP it is servicing!