Kernel: Gate architecture-specific code behind a build-time setting
This is a refactor of the existing kernel build to gate any potential
architecture-specific code behind a build-time config. The build now
expects to be passed a `MU_MACHINE` variable that indicates which
platform should be built. Based on this, a compatible toolchain is
selected (instead of manually passing the toolchains, as it happened
previously).
This could open the door to potentially porting the kernel to different
archs in the future, but for now this was done as a way of cleaning
up the existing build scripts.
There's still a lot of architecture-specific things left in the main
kernel target, but ideally they'll all be moved to the corresponding
arch modules in the future.