Kernel: Rework multicore initialization
This oversized commit reworks multicore initialization. This is now
done as part of late-stage platform initialization. The interface
for accessing the current core's control block was reworked, and is
now abstracted away from using the x86 implementation directly.
As this modifies pretty much everywhere the scheduler was used,
I'm not going to bother splitting this into separate commits.