Core: Introduce GFP / Get Free Pages subsystem
The subsystem does exactly what you'd expect - it allocates
physical memory pages for use by the kernel. GFP is supposed
to be relatively freestanding, so it is possible to use safely
immediately after boot. The only requirement is that the memory
layout was provided via core::mem::layout beforehand and
identity mapping functions are provided by the platform.