Kernel: Migrate to GFP
Removed remains of the old PMM subsystem and replaced all usages
with core::mem::gfp. x86 platform entry now also parses the GRUB
provided memory map as soon as possible during early init, so it
will be possible to use physical memory allocations pretty much
immediately after the kernel boots. Additionally, the concept of
deferred initialization was removed entirely, as it no longer is
required as we're not initializing allocators for regions that
we may not be able to access anymore.