x86_64: Bring back V86 after GFP introduction
V86 is a huge hack that should have never seen the light
of day, but still keep it alive by fixing it up enough to
work after GFP introduction. Don't mistake this V86 with
VM86 feature of the CPU though - these are not related at
all (we can't use VM86 from long-mode anyway).