Arch: VM: Return pointer of same type from idmap/idunmap expression
The idunmap fix is insufficient, and the same issue now appeared in
idmap. However, simply reinterpret_cast'ing to uint8_t is not the way to
go around this, as the physptr may be c- and v-qualified, which will
break as this cast would then discard the qualification.

Use the decltype of the input pointer instead as the result type of the
expression.