All: Remove Daemons
It is time to start fresh, remove "daemons" (actually kernel threads in
disguise) that use x86-specific bits (so all of them).

This also allows removing the ad-hoc IDE driver, which was more of a
proof of concept and shouldn't be used with threading as it uses direct
x86 I/O access.