Arch: riscv64: Add boot0 kernel bootloader
The boot0 kernel bootloader serves the same purpose as the preloader on
x86_64 - it configures barebones virtual memory so that the vmkernel can
actually run and bootstrap itself. boot0 is expected to be passed an
FDT, which it uses to determine available memory.

Boot0 is expected to be placed in memory, followed by the actual kernel,
which will start with a kernel image header. For now two separate
bootcons are supported to facilitate debugging: generic 16550 that
probably won't work outside of VisionFive 2 and SiFive UART. Boot0 is
expected to be booted from U-boot, for example (for which we can
generate an uImage that contains both boot0 and the actual kernel
binary).