LibAllocator: ChunkAllocator: Do not wipe the free chunk when splitting
There is no point to this, as the free chunk is not actually used and we
don't care about its underlying storage. This predictably leads to awful
performance when the allocator has no chunks and a large arena is used,
which leads to LibAllocator wiping practically the entirety of the arena
on every allocation.

Time-to-userland before this change: 200ms
Time-to-userland after: 150ms