Arch: x86_64: Refactor GDT to make environment struct constinit-friendly
Bring back constinit to statically defined environment structs by
refactoring GDT handling. Constructing the TSS entry for the GDT
requires reinterpret_cast'ing its address, and as such the GDT is
partially defined at compile time and the TSS entries are generated on
GDT load.

This splits the TSS out of the GDT struct as well.