CMake: Remove global enable of stack protector
Enabling stack protector for all code built using the cross-compiler
toolchain is annoying when an executable target does not currently
enable the stack protector, but depends on a static library that is
built independently. The library would be built with the stack protector,
which would force the target to also include the stack protector
symbols.

Remove stack protector configuration from global toolchain
configuration so it is possible to have the executable link against a
library that doesn't enforce use of the stack protector.